Fix bug with doubled inputs. Only use Pygame to receive keyboard inputs
Instead of a thread that listens to keyboard inputs, using the import keyboard module, all keyboard strokes now go through pygame. This means there are no longer 2 places where keyboard inputs could be entered, and this also removes the requirement for sudo rights on Unix systems. Inputs can only be entered when the Python program is on the foreground.